+++ server/v2/core/go.mod
@@ -6,22 +6,11 @@ replace cosmossdk.io/core => ../../../core
 
 require (
 	cosmossdk.io/core v0.11.0
-	cosmossdk.io/x/tx v0.13.0
 	github.com/cosmos/gogoproto v1.4.11
 	google.golang.org/protobuf v1.31.0
 )
 
 require (
-	cosmossdk.io/api v0.7.2 // indirect
-	github.com/cosmos/cosmos-proto v1.0.0-beta.3 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
 	golang.org/x/exp v0.0.0-20231006140011-7918f672742d // indirect
-	golang.org/x/net v0.19.0 // indirect
-	golang.org/x/sys v0.15.0 // indirect
-	golang.org/x/text v0.14.0 // indirect
-	google.golang.org/genproto v0.0.0-20231211222908-989df2bf70f3 // indirect
-	google.golang.org/genproto/googleapis/api v0.0.0-20231120223509-83a465c0220f // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20231212172506-995d672761c0 // indirect
-	google.golang.org/grpc v1.60.1 // indirect
 )
